+++ kiroku-store/test/Main.hs
@@ -34,6 +34,7 @@ import Test.FailureInjection qualified as FailureInjection
 import Test.Helpers
 import Test.Hspec
 import Test.InterpreterHooks qualified as InterpreterHooks
+import Test.OriginalStreamName qualified as OriginalStreamName
 import Test.Properties qualified as Properties
 import Test.ReadStream qualified as ReadStream
 import Test.Transaction qualified as Transaction
@@ -45,6 +46,7 @@ main = hspec $ do
     FailureInjection.spec
     Transaction.spec
     ReadStream.spec
+    OriginalStreamName.spec
     InterpreterHooks.spec
     Causation.spec
     ConsumerGroupSql.spec